requirements/core.txt
gluoncv == 0.6.*
pillow == 8.*
matplotlib == 3.*
numpy == 1.*
opencv-contrib-python == 4.1.*  # latest working prebuilt on rpi (via piwheels) is 4.1.0.25 (@ 2020-08-31)
scipy == 1.*
boto3 == 1.17.*
botocore == 1.20.*
